:recycle: Refs. #720 -- use @rjsf wrapper component to display validation errors
