+++ src/openforms/js/components/admin/RJSFWrapper.js
@@ -1,0 +1,59 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Form from '@rjsf/core';
+
+import Field from './forms/Field';
+
+
+const FormRjsfWrapper = ({ name, label, schema, formData, onChange, errors }) => {
+    let extraErrors = {};
+
+    // add backend validation errors in the correct format. RJSF takes nested objects,
+    // even for array types.
+    for (const [key, msg] of errors) {
+        const bits = key.split('.');
+        // create the nested structure. we can't use lodash, since it creates arrays for
+        // indices rather than nested objects.
+        let errObj = extraErrors;
+        for (const pathBit of bits) {
+            if (pathBit === '__proto__' || pathBit === 'prototype') throw new Error('Prototype polution!');
+            if (!errObj[pathBit]) errObj[pathBit] = {};
+            errObj = errObj[pathBit];
+        }
+        if (!errObj.__errors) errObj.__errors = [];
+        errObj.__errors.push(msg);
+    }
+
+    return (
+        <Field
+            name={name}
+            label={label}
+            errors={extraErrors ?  [] : errors}
+        >
+            <Form
+                schema={schema}
+                formData={formData}
+                onChange={onChange}
+                children={true}
+                extraErrors={extraErrors}
+                showErrorList={false}
+            />
+        </Field>
+    );
+};
+
+FormRjsfWrapper.propTypes = {
+    name: PropTypes.string.isRequired,
+    label: PropTypes.node.isRequired,
+    schema: PropTypes.shape({
+        type: PropTypes.oneOf(['object']), // it's the JSON schema root, it has to be
+        properties: PropTypes.object,
+        required: PropTypes.arrayOf(PropTypes.string),
+    }),
+    formData: PropTypes.object,
+    onChange: PropTypes.func.isRequired,
+    errors: PropTypes.array,
+};
+
+
+export default FormRjsfWrapper;
